requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
Pygments==2.16.1
Jinja2==3.1.2
Werkzeug==2.3.7